+++ eslint-configs/eslint-config-seekingalpha-react/rules/eslint-plugin-react/index.js
@@ -5,6 +5,15 @@ module.exports = {
   extends: [
     './react.js',
     './jsx.js'
-  ]
+  ],
+
+  settings: {
+
+    react: {
+      pragma: 'React',
+      version: '16.4.1'
+    }
+
+  }
 
 };